+++ app/scripts/api.js
@@ -666,6 +666,7 @@ const createApi = function api(context, pubSub) {
        * hgv.off('mouseMoveZoom', mmz);
        * hgv.off('wheel', wheelListener);
        * hgv.off('createSVG');
+       * hgv.off('geneSearch', geneSearchListener);
        */
       off(event, listenerId, viewId) {
         const callback =
@@ -702,6 +703,10 @@ const createApi = function api(context, pubSub) {
 
           case 'createSVG':
             self.offPostCreateSVG();
+            break;
+
+          case 'geneSearch':
+            apiPubSub.unsubscribe('geneSearch', callback);
             break;
 
           default:
@@ -875,6 +880,13 @@ const createApi = function api(context, pubSub) {
        *    svg.appendChild(circle);
        *    return svg;
        * });
+       *
+       * const geneSearchListener = event => {
+       *    console.log('Gene searched', event.geneSymbol);
+       *    console.log('Range of the gene', event.range);
+       *    console.log('Center of the gene', event.centerX);
+       * }
+       * hgv.on('geneSearch', geneSearchListener);
        */
       on(event, callback, viewId, callbackId) {
         switch (event) {
@@ -902,6 +914,9 @@ const createApi = function api(context, pubSub) {
 
           case 'createSVG':
             return self.onPostCreateSVG(callback);
+
+          case 'geneSearch':
+            return apiPubSub.subscribe('geneSearch', callback);
 
           default:
             return undefined;