+++ test/APITests.js
@@ -975,4 +975,26 @@ describe('API Tests', () => {
       div = undefined;
     });
   });
+
+  it('triggers on gene search events', done => {
+    [div, api] = createElementAndApi(simple1dHorizontalVerticalAnd2dDataTrack);
+    const hgc = api.getComponent();
+    waitForTilesLoaded(hgc, () => {
+      api.on('geneSearch', e => {
+        expect(e.geneSymbol).toEqual('MYC');
+        expect(e.centerX).toEqual(1521546687);
+        done();
+      });
+
+      // The gene search event that we expect to catch.
+      const geneSearchEvent = {
+        geneSymbol: 'MYC',
+        range: [1521542663, 1521550711],
+        centerX: 1521546687,
+        centerY: 1521546687,
+      };
+      // Simulate the gene search events.
+      hgc.geneSearchHandler(geneSearchEvent);
+    });
+  });
 });